calendar-assistant/calendar-assistant/src/main/java/com/calendarassistant/calendar_assistant/controller/MeetingController.java
package com.calendarassistant.calendar_assistant.controller;

import com.calendarassistant.calendar_assistant.exception.ConflictException;
import com.calendarassistant.calendar_assistant.exception.NotFoundException;
import com.calendarassistant.calendar_assistant.modal.Meeting;
import com.calendarassistant.calendar_assistant.service.MeetingService;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.*;

import java.util.List;

@RestController
@RequestMapping("/api/meetings")
public class MeetingController {

    private final MeetingService meetingService;

    @Autowired
    public MeetingController(MeetingService meetingService) {
        this.meetingService = meetingService;
    }

    @GetMapping("/{id}")
    public ResponseEntity<Meeting> getMeetingById(@PathVariable Long id) {
        try {
            Meeting meeting = meetingService.findById(id);
            return ResponseEntity.ok(meeting);
        } catch (NotFoundException e) {
            return ResponseEntity.notFound().build();
        }
    }

    @PostMapping
    public ResponseEntity<Meeting> createMeeting(@RequestBody Meeting meeting) {
        try {
            Meeting savedMeeting = meetingService.save(meeting);
            return ResponseEntity.ok(savedMeeting);
        } catch (ConflictException e) {
            return ResponseEntity.status(409).build();
        }
    }

    @DeleteMapping("/{id}")
    public ResponseEntity<Void> deleteMeeting(@PathVariable Long id) {
        try {
            meetingService.deleteById(id);
            return ResponseEntity.noContent().build();
        } catch (NotFoundException e) {
            return ResponseEntity.notFound().build();
        }
    }

    @GetMapping
    public ResponseEntity<List<Meeting>> getAllMeetings() {
        List<Meeting> meetings = meetingService.findAll();
        return ResponseEntity.ok(meetings);
    }
}
